change int to size_t where needed + list clone impl

+++ include/object.h
@@ -6,5 +6,5 @@ typedef struct Object Object;
 void Object_ctor(Object* obj, char* name);
 void Object_dtor(Object* obj);
 char* Object_toString(Object* object);
-int Object_hashCode(Object* object);
+size_t Object_hashCode(Object* object);
 Object* Object_clone(Object* object);
